here comes -stable

+++ sys/dev/rcons/raster.h
@@ -1,4 +1,5 @@
-/*	$NetBSD: raster.h,v 1.2 1995/10/04 23:57:19 pk Exp $ */
+/*	$OpenBSD: raster.h,v 1.4 2002/03/14 03:16:07 millert Exp $ */
+/*	$NetBSD: raster.h,v 1.3 1995/11/24 23:50:51 cgd Exp $ */
 
 /*-
  * Copyright (c) 1991, 1993
@@ -16,11 +17,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *	This product includes software developed by the University of
- *	California, Berkeley and its contributors.
- * 4. Neither the name of the University nor the names of its contributors
+ * 3. Neither the name of the University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
  *
@@ -87,16 +84,16 @@ struct raster {
     int width, height;	/* size in pixels */
     int depth;		/* bits per pixel - 1 or 8 */
     int linelongs;	/* longs from one line to the next - for padding */
-    u_long* pixels;	/* pointer to the actual bits */
+    u_int32_t *pixels;	/* pointer to the actual bits */
     caddr_t data;	/* special pointer for frame buffers and subregions */
     };
 
 /* Colormap struct. */
 struct raster_colormap {
     int length;
-    u_char* red;
-    u_char* grn;
-    u_char* blu;
+    u_char *red;
+    u_char *grn;
+    u_char *blu;
     };
 
 /* Font character struct. */
@@ -209,13 +206,13 @@ extern int raster_replsrc ARGS(( struct raster* dst, int dx, int dy, int w, int 
 
 /* Raster text routines */
 
-extern struct raster_font* raster_fontopen ARGS(( char* fontname ));
+extern struct raster_font* raster_fontopen ARGS(( char *fontname ));
 /* Opens a font. Returns (struct raster_font*) 0 on failure. */
 
-extern int raster_text ARGS(( struct raster* r, int x, int y, int rop, struct raster_font* rf, unsigned char* text ));
+extern int raster_text ARGS(( struct raster* r, int x, int y, int rop, struct raster_font* rf, unsigned char *text ));
 /* Draws text.  Returns 0 on success, -1 on failure. */
 
-extern int raster_textn ARGS(( struct raster* r, int x, int y, int rop, struct raster_font* rf, unsigned char* text, int len ));
+extern int raster_textn ARGS(( struct raster* r, int x, int y, int rop, struct raster_font* rf, unsigned char *text, int len ));
 /* Draws n characters of text.  Returns 0 on success, -1 on failure. */
 
 extern void raster_fontclose ARGS(( struct raster_font* rf ));
@@ -224,7 +221,7 @@ extern void raster_fontclose ARGS(( struct raster_font* rf ));
 
 /* Frame buffer routines. */
 
-extern struct raster* raster_open ARGS(( char* fbname ));
+extern struct raster* raster_open ARGS(( char *fbname ));
 /* Opens a frame buffer as a raster.  Returns (struct raster*) 0 on failure. */
 
 extern struct raster* raster_coloropen ARGS(( void ));
